qt: Fix deprecated QCharRef usage
Port https://github.com/bitcoin/bitcoin/pull/18101 to
our codebase. Without this commit every Alt-Tab or other
keypress on the SplashScreen will lead to the following
message in debug.log:

GUI: Using QCharRef with an index pointing outside the valid range of a QString. The corresponding behavior is deprecated, and will be changed in a future version of Qt.