fix libsnark build with optimisations on gcc/g++ 11.x
It always was default behavior to build KomodoOcean with `-DUSE_ASM=1`
which is set by default through `config/bitcoin-config.h`. This flag
used not only by `libsnark`, but also by the hardrware SHA256 optimisations
in other sources, etc. However, it turns out that on g++ 11.3.0 enabled
optimisations for libsnark makes ibzcash::PHGRProof (SproutProofVerifier)
not working, and as a result when users tried to sync old chains which
had `sprout` transactions in history, like `SUPERNET`, the verification
of valid transactions throw an error:

```
2023-07-11 15:01:07 ERROR: CheckTransaction(): joinsplit does not verify
2023-07-11 15:01:07 ERROR: CheckBlock: CheckTransaction failed
```

While on gcc/g++ 10.x there was no such error. That's why we decided to
disable `libsnark` optimisations at all, by "shadowing" `USE_ASM` in
internal libnark implementation and rename it to `USE_ASM_SNARK`, to
avoid build with ASM optimisations, which are "unstable" on modern compilers.

More details here:

https://github.com/KomodoPlatform/komodo/issues/591