tests: fix compatibility of LegacyEvents test with other tests
For example, the group of test_block tests changes the value
of IsInitialBlockDownload() to false after the first call
to generateBlock. As a result, when komodo_connectblock is called,
komodo_validate_chain should return 1 for the LegacyEvents tests.
Additionally, if we are operating with various data directories
between different tests, the static file pointer fp in komodo.cpp
should be cleared between these tests to avoid writing to the same
file in different tests.