strip cutesy branding
"komodo" this, that, and the other is great if you want to sell a thing; not so much for devs that want to work on things and import patches and upgrades from around the larger crypto ecosystem.
RE: komodo/bitcoinstrings.cpp - willing to manually resolve strings conflicts as needed, but not that AND file name issues at the same time.
This can all be reversed AFTER the products are brought into this decade.
(survives make clean and make - will address any missed items if found)